[refactor] Watch 리팩토링 및 주석 추가
