+++ server/src/main/java/com/moovda_project/moovda/module/comment/entity/Comment.java
@@ -7,9 +7,6 @@ import com.moovda_project.moovda.module.movie.entity.Movie;
 import lombok.Getter;
 import lombok.NoArgsConstructor;
 import lombok.Setter;
-import org.hibernate.annotations.OnDelete;
-import org.hibernate.annotations.OnDeleteAction;
-
 import javax.persistence.*;
 import java.util.HashSet;
 import java.util.Set;
