+++ server/src/main/java/com/moovda_project/moovda/module/member/controller/MemberController.java
@@ -44,10 +44,10 @@ public class MemberController {
 
     @GetMapping("{member_id}")
     public ResponseEntity getMember(@PathVariable("member_id") @Positive long memberId,
-                                    @Positive @RequestParam int page1,
-                                    @Positive @RequestParam int page2) {
+                                    @Positive @RequestParam int toWatchPage,
+                                    @Positive @RequestParam int watchedPage) {
         Member member = memberService.findMember(memberId);
 
-        return new ResponseEntity<>(mapper.memberToMemberResponseDto(member,page1,page2,5), HttpStatus.OK);
+        return new ResponseEntity<>(mapper.memberToMemberResponseDto(member,toWatchPage,watchedPage,5), HttpStatus.OK);
     }
 }
