+++ server/src/main/java/com/moovda_project/moovda/module/member/dto/MemberResponseDto.java
@@ -1,12 +1,10 @@
 package com.moovda_project.moovda.module.member.dto;
 
-import com.moovda_project.moovda.module.movie.dto.PageDto;
+import com.moovda_project.moovda.global.dto.PageDto;
 import com.moovda_project.moovda.module.watch.dto.ToWatchResponseDto;
 import com.moovda_project.moovda.module.watch.dto.WatchedResponseDto;
-import lombok.Builder;
 import lombok.Getter;
 import lombok.Setter;
-import org.springframework.data.domain.Page;
 
 import java.util.List;
