+++ server/src/main/java/com/moovda_project/moovda/module/watch/repository/ToWatchRepository.java
@@ -9,6 +9,5 @@ import java.util.Optional;
 
 public interface ToWatchRepository extends JpaRepository<ToWatch, Long> {
     Optional<ToWatch> findByMemberAndMovie(Member member, Movie movie);
-
     void deleteByMemberAndMovie(Member member, Movie movie);
 }
